requirements.txt
flask==3.0.0
gunicorn==21.2.0
pillow==10.1.0
requests==2.31.0
numpy==1.24.3
onnxruntime==1.16.0